+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Command-line option `--init` to create a new changelog if it does not exist.
 - Command-line option `--check` to check if changelog has any changes under \[Unreleased\].
 
+### Removed
+- **[BREAKING CHANGE]** Removed support for Node versions <= 8
+
 ## [1.1.0] - 2018-01-02
 ### Added
 - Support for BitBucket