+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.0.0] - 2020-09-23
 ### Added
 - Command-line option `--init` to create a new changelog if it does not exist.
 - Command-line option `--check` to check if changelog has any changes under \[Unreleased\].
@@ -25,6 +27,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Initial commit
 
-[Unreleased]: https://github.com/nikolajevp/changelog-updater/compare/v1.1.0...HEAD
+[Unreleased]: https://github.com/nikolajevp/changelog-updater/compare/v2.0.0...HEAD
+[2.0.0]: https://github.com/nikolajevp/changelog-updater/compare/v1.1.0...v2.0.0
 [1.1.0]: https://github.com/nikolajevp/changelog-updater/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/nikolajevp/changelog-updater/compare/v0.1.0...v1.0.0